kafka-cluster/python-client/requirements.txt
confluent-kafka==0.11.5
kafka-python==1.4.3
